Split the CloudSearch handling into a separate service
